Gene/requirements.txt
# Gene SQL Query System Requirements

# MySQL Database Connection
mysql-connector-python>=8.0.0

# HTTP Requests for AI Gateway
requests>=2.25.0

# Environment Variables (if needed)
python-dotenv>=0.19.0

# Optional: For advanced SQL parsing
sqlparse>=0.4.0